+++ man-roxygen/param-loss.R
@@ -1,0 +1,3 @@
+#' @param loss ([LossQuadratic] | [LossBinomial] | [LossHuber] | [LossAbsolute] | [LossQuantile])\cr
+#' An initialized `S4` loss object (requires to call `Loss*$new(...)`).
+#' See the respective help page for further information.