+++ src/baselearner_factory.h
@@ -122,12 +122,13 @@ class BaselearnerPolynomialFactory : public BaselearnerFactory
 {
 private:
   sbindata _sh_ptr_bindata;
-  const std::shared_ptr<init::PolynomialAttributes> _attributes = std::make_shared<init::PolynomialAttributes>();
 
 public:
   BaselearnerPolynomialFactory (const std::string, std::shared_ptr<data::Data>,
     const unsigned int, const bool, const unsigned int, const double = 0, const double = 0);
   BaselearnerPolynomialFactory (const json&, const mdata&, const mdata&);
+
+  std::shared_ptr<init::PolynomialAttributes> _attributes = std::make_shared<init::PolynomialAttributes>();
 
   bool       usesSparse           ()                 const;
   sdata      instantiateData      (const mdata&)     const;
@@ -154,13 +155,14 @@ class BaselearnerPSplineFactory : public BaselearnerFactory
 {
 private:
   sbindata _sh_ptr_bindata;
-  std::shared_ptr<init::PSplineAttributes> _attributes = std::make_shared<init::PSplineAttributes>();
 
 public:
   BaselearnerPSplineFactory (const std::string, const std::shared_ptr<data::Data>&, const unsigned int,
     const unsigned int, const double, const double, const unsigned int, const bool, const unsigned int,
     const std::string);
   BaselearnerPSplineFactory (const json&, const mdata&, const mdata&);
+
+  std::shared_ptr<init::PSplineAttributes> _attributes = std::make_shared<init::PSplineAttributes>();
 
   bool       usesSparse           ()                 const;
   sdata      instantiateData      (const mdata&)     const;
@@ -209,6 +211,9 @@ public:
   arma::vec  getPenalty           ()                 const;
   arma::mat  getPenaltyMat        ()                 const;
 
+  std::shared_ptr<blearnerfactory::BaselearnerFactory> getBl1 () const;
+  std::shared_ptr<blearnerfactory::BaselearnerFactory> getBl2 () const;
+
   arma::mat  calculateLinearPredictor (const arma::mat&) const;
   arma::mat  calculateLinearPredictor (const arma::mat&, const mdata&) const;
 
@@ -226,7 +231,6 @@ public:
 class BaselearnerCenteredFactory : public BaselearnerFactory
 {
 private:
-  std::shared_ptr<init::CenteredAttributes> _attributes = std::make_shared<init::CenteredAttributes>();
   std::shared_ptr<data::BinnedData> _sh_ptr_bindata;
 
   std::shared_ptr<blearnerfactory::BaselearnerFactory> _blearner1;
@@ -236,6 +240,8 @@ public:
   BaselearnerCenteredFactory (const std::string&, std::shared_ptr<blearnerfactory::BaselearnerFactory>,
     std::shared_ptr<blearnerfactory::BaselearnerFactory>);
   BaselearnerCenteredFactory (const json&, const mdata&, const mdata&);
+
+  std::shared_ptr<init::CenteredAttributes> _attributes = std::make_shared<init::CenteredAttributes>();
 
   bool       usesSparse           ()                 const;
   sdata      instantiateData      (const mdata&)     const;
@@ -264,9 +270,10 @@ class BaselearnerCategoricalRidgeFactory : public BaselearnerFactory
 {
 private:
   sdata _sh_ptr_data;
-  std::shared_ptr<init::RidgeAttributes> _attributes = std::make_shared<init::RidgeAttributes>();
 
 public:
+  std::shared_ptr<init::RidgeAttributes> _attributes = std::make_shared<init::RidgeAttributes>();
+
   BaselearnerCategoricalRidgeFactory (const std::string, std::shared_ptr<data::CategoricalDataRaw>&, const double = 0, const double = 0);
   BaselearnerCategoricalRidgeFactory (const json&, const mdata&, const mdata&);
 
@@ -299,9 +306,10 @@ class BaselearnerCategoricalBinaryFactory : public BaselearnerFactory
 {
 private:
   sdata _sh_ptr_data;
-  std::shared_ptr<init::BinaryAttributes> _attributes = std::make_shared<init::BinaryAttributes>();
 
 public:
+  std::shared_ptr<init::BinaryAttributes> _attributes = std::make_shared<init::BinaryAttributes>();
+
   BaselearnerCategoricalBinaryFactory (const std::string, const std::string, const std::shared_ptr<data::CategoricalDataRaw>&);
   BaselearnerCategoricalBinaryFactory (const json&, const mdata&, const mdata&);