+++ NEWS.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+- **21.12.2022** \
+  New methods for transforming data are now available (`cboost$transformData(newdat)`). Additionally, methods for accessing the meta data of a base learner were also added (`cboost$baselearner_list$blfactory$factory$getMeta()`).
+
+- **09.12.2022** \
+  Re-writing the documentation for the `Compboost` class.
+
 - **20.04.2022** \
   Adding a new "intercept base learner". This base learner can be used to additionally add an intercept. This makes sense if, e.g. linear functions without intercept are added.